Add files via upload
Got rid of (!Serial) lines to make sure the device runs the script uploaded to it automatically when plugged into power. This line means that the device is waiting for the serial monitor to start up before running the script. The serial monitor only starts up when the device is plugged into a laptop and the device needs to run the script automatically when powered by any source, not just a laptop.